[FIX] website: display back the "Position" grid image option
Since commit [1], the "Position" grid image option is not appearing
anymore, making it impossible to change the `object-fit` property of the
grid item images, making them forced to always be in `cover`.

The fix of commit [1] was made to avoid a traceback when toggling back
to normal mode, because the `isApplied` was run after the grid mode was
disabled but before the option noticed that it was not supposed to be
there anymore. However, this was not the right fix, as it made the
option not appear at all.

This commit removes this previous fix and fixes the issue by simply
checking if the `o_grid_item_image` element is well present in
`isApplied`(as it was done before the refactoring).

This commit also cleans the code a bit and renames the files/component/
plugin/action so they are called `grid_image_` instead of `image_grid_`,
to be more correct and to make the link with the old code.

[1]: 82882b7e3e24951aa7224a3a266e57cc4773e546

task-4367641
