--- addons/website/static/src/builder/plugins/image/image_grid_option.xml
@@ -1,13 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<templates xml:space="preserve">
-
-<t t-name="website.ImageGridOption">
-    <BuilderRow t-if="isActiveItem('grid_mode') and state.isOptionActive" label.translate="Position">
-        <BuilderSelect action="'setGridImageMode'" >
-            <BuilderSelectItem title.translate="Contain" actionValue="'contain'">Contain</BuilderSelectItem>
-            <BuilderSelectItem title.translate="Cover" actionValue="'cover'">Cover</BuilderSelectItem>
-        </BuilderSelect>
-    </BuilderRow>
-</t>
-
-</templates>
